DKMS requires executing make
Even it might be possible to force DKMS executing a script instead of make
it seems to be much easier to add a Makefile for DKMS and execute
build_all.sh from within this new Makefile.dkms.

dkms.conf will now execute the new Makefile.dkms for build and clean.