media_build: Added missing functions nsecs_to_jiffies(64)
Several modules expect the functions nsecs_to_jiffies64 and
nsecs_to_jiffies to be available when they get loaded. For Kernels prior
to 3.16, this symbol is not exported in time.c .
Copied the function to compat.h, so that it gets already resolved during
compilation. Define also a macro with a name conversion, because the
mentioned functions are defined as extern in include/linux/jiffies.h,
which gives an error when the are re-defined as static.

Signed-off-by: Jasmin Jessich <jasmin@anw.at>